requirements.txt
faiss-cpu
ffmpeg-python
numpy
openai-whisper
pandas
pydub
requests
sentence_transformers
srt
streamlit
tabulate
torch
tqdm
transformers
yt_dlp
